+++ config/connection.js
@@ -1,7 +1,7 @@
 const Sequelize = require('sequelize');
 require('dotenv').config()
-const session = require('express-session');
-const SequelizeStore = require('connect-session-sequelize')(session.Store);
+// const session = require('express-session');
+// const SequelizeStore = require('connect-session-sequelize')(session.Store);
 
 
 let sequelize;
@@ -21,23 +21,25 @@ if (process.env.JAWSDB_URL) {
   );
 }
 
-const sess = {
-  secret: 'Super goat secret',
+// const sess = {
+//   secret: 'Super goat secret',
   
-  cookie: {
-    maxAge: 24 * 60 * 60 * 1000,
-    httpOnly: true,
-    secure: false,
-    sameSite: 'strict',
-  },
+//   cookie: {
+//     maxAge: 24 * 60 * 60 * 1000,
+//     httpOnly: true,
+//     secure: false,
+//     sameSite: 'strict',
+//   },
 
-  resave: false,
-  saveUninitialized: true,
-  store: new SequelizeStore({
-    db: sequelize
-  })
-};
+//   resave: false,
+//   saveUninitialized: true,
+//   store: new SequelizeStore({
+//     db: sequelize
+//   })
+// };
 
-const sessionMiddleware = session(sess);
+// const sessionMiddleware = session(sess);
 
-module.exports = {sequelize,sessionMiddleware};
+module.exports = sequelize;
+
+  // ,sessionMiddleware};
