+++ server.js
@@ -4,7 +4,7 @@ const session = require('express-session');
 const exphbs = require('express-handlebars');
 const routes = require('./controllers');
 const sequelize = require('./config/connection');
-const {sessionMiddleware} = require('./config/connection');
+// const {sessionMiddleware} = require('./config/connection');
 const SequelizeStore = require('connect-session-sequelize')(session.Store);
 
 const app = express();
@@ -12,7 +12,24 @@ const PORT = process.env.PORT || 3001;
 
 const hbs = exphbs.create();
 
-app.use(sessionMiddleware);
+const sess = {
+  secret: 'Super goat secret',
+  
+  cookie: {
+    maxAge: 24 * 60 * 60 * 1000,
+    httpOnly: true,
+    secure: false,
+    sameSite: 'strict',
+  },
+
+  resave: false,
+  saveUninitialized: true,
+  store: new SequelizeStore({
+    db: sequelize
+  })
+};
+
+app.use(sess);
 
 app.engine('handlebars', hbs.engine);
 app.set('view engine', 'handlebars');
